Added : Employee CRUD

+++ api/employee/views.py
@@ -11,22 +11,28 @@ from rest_framework.permissions import IsAuthenticated
 
 class EmployeeView(views.APIView):
     permission_classes = [IsAuthenticated]
-    def get(self,request,organization_id):
+    def get(self,request,organization_id,department_id):
         try:
             org = Organization.objects.filter(organization_id=organization_id,user=request.user).first()
             if org == None:
                return Response({"status": "success", "message":"Invalid organization"}, status=status.HTTP_400_OK) 
-            depts = Department.objects.filter(organization=org)
-            serializer = DepartmentSerializer(depts,many=True,context={'user':request.user})
+            dept = Department.objects.filter(department_id=department_id,organization=org).first()
+            if dept == None:
+               return Response({"status": "success", "message":"Invalid department"}, status=status.HTTP_400_OK)
+            empls = Employee.objects.filter(department=dept)
+            serializer = EmployeeSerializer(empls,many=True,context={'user':request.user})
             return Response({"status": "success", "data": serializer.data}, status=status.HTTP_200_OK) 
         except Exception as e:
             return Response({"status":"internal-error","message":"An unexpected error occured"})
-    def post(self, request,organization_id):
+    def post(self, request,organization_id,department_id):
       try:
         org = Organization.objects.filter(organization_id=organization_id,user=request.user).first()
         if org == None:
-            return Response({"status": "success", "message":"Invalid organization"}, status=status.HTTP_400_OK)
-        serializer = DepartmentSerializer(data=request.data,context={'organization':org})
+            return Response({"status": "success", "message":"Invalid organization"}, status=status.HTTP_400_OK) 
+        dept = Department.objects.filter(department_id=department_id,organization=org).first()
+        if dept == None:
+            return Response({"status": "success", "message":"Invalid department"}, status=status.HTTP_400_OK)
+        serializer = EmployeeSerializer(data=request.data,context={'department':dept})
         if serializer.is_valid():  
             serializer.save() 
             return Response({"status": "success", "data": serializer.data}, status=status.HTTP_200_OK)  
@@ -35,16 +41,19 @@ class EmployeeView(views.APIView):
       except Exception as e:
           return Response({"status":"internal-error","message":"An unexpected error occured"})
 
-    def put(self, request,organization_id):
+    def put(self, request,organization_id,department_id):
       try:
         org = Organization.objects.filter(organization_id=organization_id,user=request.user).first()
         if org == None:
             return Response({"status": "success", "message":"Invalid organization"}, status=status.HTTP_400_OK)
-        instance = Department.objects.filter(department_id = request.data.get('department_id'),organization=org).first()
+        dept = Department.objects.filter(department_id=department_id,organization=org).first()
+        if dept == None:
+            return Response({"status": "success", "message":"Invalid department"}, status=status.HTTP_400_OK)
+        instance = Employee.objects.filter(employee_id = request.data.get('employee_id'),department=dept).first()
         if instance == None:
-           return Response({'status':'error','message':'invalid org id'})
+           return Response({'status':'error','message':'invalid employee id'})
         
-        serializer = DepartmentUpdateSerializer(instance,data=request.data) 
+        serializer = EmployeeUpdateSerializer(instance,data=request.data)
         if serializer.is_valid():  
             serializer.save() 
             return Response({"status": "success", "data": serializer.data}, status=status.HTTP_200_OK)  
@@ -54,19 +63,19 @@ class EmployeeView(views.APIView):
           return Response({"status":"internal-error","message":"An unexpected error occured"})
 
 
-    def delete(self, request,organization_id):
+    def delete(self, request,organization_id,department_id):
       try:
         org = Organization.objects.filter(organization_id=organization_id,user=request.user).first()
         if org == None:
             return Response({"status": "success", "message":"Invalid organization"}, status=status.HTTP_400_OK)
-        instance = Department.objects.filter(department_id = request.data.get('department_id'),organization=org).first()
+        dept = Department.objects.filter(department_id=department_id,organization=org).first()
+        if dept == None:
+            return Response({"status": "success", "message":"Invalid department"}, status=status.HTTP_400_OK)
+        instance = Employee.objects.filter(employee_id = request.data.get('employee_id'),department=dept).first()
         if instance == None:
-           return Response({'status':'error','message':'invalid dept id'})
+           return Response({'status':'error','message':'invalid employee id'})
         instance.delete()
-        return Response({"status": "success", "message":"successfully deleted organization"}, status=status.HTTP_200_OK) 
+        return Response({"status": "success", "message":"successfully deleted employee"}, status=status.HTTP_200_OK) 
       except Exception as e:
           return Response({"status":"internal-error","message":"An unexpected error occured"})
     
-    
-        
-
